+++ app/profile/[id]/page.tsx
@@ -2,30 +2,22 @@
 import { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { RootState, AppDispatch } from '@/src/store/store';
-import { updateUserImage, updateUserInfo, uploadImage } from '@/src/store/auth/userThunks';
-import { useRouter, useParams } from 'next/navigation';
-import LoadPage from '@/src/components/ui/loadpage';
-import { fetchUserById } from '@/src/store/userSlice';
+import { useParams } from 'next/navigation';
+
+import { fetchUserById } from '@/src/store/slice/userSlice';
 
 export default function UserProfile() {
     const [userDetails, setUserDetails] = useState({
         name: '',
         email: '',
         address: '',
-        picture: '', // Field for image_url
+        picture: '',
     });
-    // const [loading, setLoading] = useState(true);
-    const [newPassword, setNewPassword] = useState('');
-    const [confirmPassword, setConfirmPassword] = useState('');
-    // const dispatch = useAppDispatch();
-    const router = useRouter();
-    const { id } = useParams(); // Get user ID from route params
-    const [errors, setErrors] = useState('');
-    const [successMessage, setSuccessMessage] = useState('');
-    const [imagePreview, setImagePreview] = useState<string | null>(null);
-    const [uploadButton, setUploadButton] = useState<boolean>(false);
+
+    const { id } = useParams();
+    const [errors] = useState('');
     const dispatch = useDispatch<AppDispatch>();
-    const { user, loading, error } = useSelector((state: RootState) => state.user);
+    const { user } = useSelector((state: RootState) => state.user);
     useEffect(() => {
         if (!id) return;
 
@@ -34,47 +26,6 @@ export default function UserProfile() {
 
         dispatch(fetchUserById(numericId)).unwrap();
     }, [dispatch, id]);
-
-    const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const file = e.target.files?.[0];
-        if (file) {
-            const reader = new FileReader();
-            reader.onloadend = () => {
-                setImagePreview(reader.result as string);
-                setUserDetails({ ...userDetails, picture: reader.result as string });
-            };
-            reader.readAsDataURL(file);
-        }
-        setUploadButton(true);
-    };
-
-    const handleUpload = async (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        const formData = new FormData(event.currentTarget);
-
-        try {
-            const resultAction = await dispatch(uploadImage(formData));
-            if (uploadImage.fulfilled.match(resultAction)) {
-                await dispatch(updateUserImage());
-            }
-            router.push('/property');
-        } catch (error) {
-            console.error('Upload error:', error);
-            alert('Upload failed!');
-        }
-    };
-
-    const handleSubmit = async (e: React.FormEvent) => {
-        e.preventDefault();
-        const resultAction = await dispatch(updateUserInfo(userDetails));
-        if (updateUserInfo.fulfilled.match(resultAction)) {
-            router.push('/property');
-        } else {
-            setErrors('Failed to update profile');
-        }
-    };
-
-    if (loading) return <LoadPage />;
 
     return (
         <div className="flex min-h-screen items-center justify-center bg-gray-100">
@@ -88,7 +39,7 @@ export default function UserProfile() {
                     </div>
                 )}
 
-                <form onSubmit={handleUpload} className="space-y-6">
+                <form className="space-y-6">
                     {/* Profile Picture */}
                     <div className="flex flex-col items-center mb-6">
                         <div className="w-32 h-32 rounded-full overflow-hidden border-4 border-gray-200">
