+++ app/users/dashboard/lessor/components/CreateNewProperty.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { Dispatch, SetStateAction, useState } from 'react';
 import { useAppDispatch } from '@/src/store/hooks';
-import { createProperty, Property } from '@/src/store/propertySlice';
+import { createProperty, Property } from '@/src/store/slice/propertySlice';
 
 type CreateNewPropertyProps = {
     setIsCreateNewPropertyVisible: Dispatch<SetStateAction<boolean>>;
