+++ app/users/dashboard/lessor/components/Table/RequestTable.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, Dispatch, SetStateAction } from 'react';
+import React, { useState, useEffect } from 'react';
 import { getRequestData, requestData } from '@/src/api/data/request';
 import Footer from '../Footer';
 import RequestSlider from '../Slider/RequestSlider';
